// insurance_Front(3)/insurance_Front(3)/src/app/models/claim.model.ts
import { OurUsers } from "./user.model";

export enum ClaimType {
    CLAIM = 'CLAIM',
    IDEAS = 'IDEAS'
  }

  export enum ClaimStatus {
    UNTREATED = 'UNTREATED',
    TREATED = 'TREATED'
}
export interface Claim{
        claimId: number;
        description: string;
        dateCreation: string; 
        claimStatus: ClaimStatus;
        claimType: ClaimType;
        user?: OurUsers
}